+++ mithril-client/src/client.rs
@@ -9,7 +9,6 @@ use mithril_common::api_version::APIVersionProvider;
 use mithril_common::{MITHRIL_CLIENT_TYPE_HEADER, MITHRIL_ORIGIN_TAG_HEADER};
 
 use crate::aggregator_client::{AggregatorClient, AggregatorHTTPClient};
-#[cfg(feature = "unstable")]
 use crate::cardano_database_client::CardanoDatabaseClient;
 use crate::cardano_stake_distribution_client::CardanoStakeDistributionClient;
 use crate::cardano_transaction_client::CardanoTransactionClient;
@@ -103,7 +102,6 @@ pub struct Client {
     certificate_client: Arc<CertificateClient>,
     mithril_stake_distribution_client: Arc<MithrilStakeDistributionClient>,
     snapshot_client: Arc<SnapshotClient>,
-    #[cfg(feature = "unstable")]
     cardano_database_client: Arc<CardanoDatabaseClient>,
     cardano_transaction_client: Arc<CardanoTransactionClient>,
     cardano_stake_distribution_client: Arc<CardanoStakeDistributionClient>,
@@ -132,7 +130,6 @@ impl Client {
     }
 
     /// Get the client that fetches and downloads Cardano database snapshots.
-    #[cfg(feature = "unstable")]
     pub fn cardano_database_v2(&self) -> Arc<CardanoDatabaseClient> {
         self.cardano_database_client.clone()
     }
@@ -289,7 +286,6 @@ impl ClientBuilder {
             logger.clone(),
         ));
 
-        #[cfg(feature = "unstable")]
         let cardano_database_client = Arc::new(CardanoDatabaseClient::new(
             aggregator_client.clone(),
             #[cfg(feature = "fs")]
@@ -312,7 +308,6 @@ impl ClientBuilder {
             certificate_client,
             mithril_stake_distribution_client,
             snapshot_client,
-            #[cfg(feature = "unstable")]
             cardano_database_client,
             cardano_transaction_client,
             cardano_stake_distribution_client,